+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changes since 1.1.2
 - feat: add wait option
 - feat: apiVersion 1.1 for wait option and add better version handling
+- fix: stop creation of revisions if the helm compose file didn't change
 
 # 1.1.2
 - chore: dependency upgrades