+++ CHANGELOG.md
@@ -1,4 +1,5 @@
-# Changes since 1.1.2
+# Changes since 1.2.0
+# 1.2.0
 - feat: add wait option
 - feat: apiVersion 1.1 for wait option and add better version handling
 - fix: stop creation of revisions if the helm compose file didn't change